refactor(mitm-addon): split compiled firewall precedence tests
